feat: crud role

+++ internal/handlers/role_handlers.go
@@ -52,7 +52,14 @@ func (h *RoleHandlers) Update(c *fiber.Ctx) error {
 	if err := c.BodyParser(&req); err != nil {
 		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Invalid role data"})
 	}
+
 	req.Id = c.Params("id")
+
+	role, err := h.roleService.GetById(c.Context(), req.Id)
+	if role.Name == "admin" {
+		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Cannot update admin role"})
+	}
+
 	res, err := h.roleService.Update(c.Context(), req.Id, req)
 	if err != nil {
 		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": err.Error()})
@@ -62,6 +69,10 @@ func (h *RoleHandlers) Update(c *fiber.Ctx) error {
 
 func (h *RoleHandlers) Delete(c *fiber.Ctx) error {
 	id := c.Params("id")
+	role, err := h.roleService.GetById(c.Context(), id)
+	if role.Name == "admin" {
+		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Cannot update admin role"})
+	}
 	res, err := h.roleService.Delete(c.Context(), id)
 	if err != nil {
 		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": fmt.Sprintf("Failed to delete role: %s", err)})
